# production/contracts/ai-openapi.yaml
...
openapi: 3.1.0
info:
  title: BhooAI Nexus AI
  version: 1.0.0
  description: OpenAI-compatible AI gateway fronting OpenAI and Ollama.
servers:
  - url: http://localhost:8000
paths:
...
  /chat/completions:
    post:
      summary: Create a chat completion (streaming or not).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChatCompletionRequest' }
      responses:
        '200':
          description: A chat completion (JSON) or an SSE stream of chunks.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatCompletionResponse' }
            text/event-stream:
              schema: { $ref: '#/components/schemas/ChatCompletionChunk' }
...
components:
  schemas:
    Provider:
      type: string
      enum: [openai, ollama, auto]
    Role:
      type: string
      enum: [system, user, assistant, tool]
    Message:
      type: object
      required: [role, content]
      properties:
        role: { $ref: '#/components/schemas/Role' }
        content: { type: string }
        name: { type: string }
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model: { type: string }
        messages:
          type: array
          items: { $ref: '#/components/schemas/Message' }
        stream: { type: boolean, default: false }
        temperature: { type: number }
        max_tokens: { type: integer }
        provider: { $ref: '#/components/schemas/Provider' }
    Choice:
      type: object
      properties:
        index: { type: integer }
        message: { $ref: '#/components/schemas/Message' }
        finish_reason: { type: string }
    ChatCompletionResponse:
      type: object
      properties:
        id: { type: string }
        model: { type: string }
        provider: { type: string }
        choices:
          type: array
          items: { $ref: '#/components/schemas/Choice' }
        usage:
          type: object
          properties:
            prompt_tokens: { type: integer }
            completion_tokens: { type: integer }
            total_tokens: { type: integer }
    Delta:
      type: object
      properties:
        role: { $ref: '#/components/schemas/Role' }
        content: { type: string }
    ChunkChoice:
      type: object
      properties:
        index: { type: integer }
        delta: { $ref: '#/components/schemas/Delta' }
        finish_reason: { type: string, nullable: true }
    ChatCompletionChunk:
      type: object
      properties:
        id: { type: string }
        model: { type: string }
        provider: { type: string }
        choices:
          type: array
          items: { $ref: '#/components/schemas/ChunkChoice' }
...